refactor: move retry functionality to utils package
