+++ azuredevops/internal/utils/retry.go
@@ -1,4 +1,4 @@
-package workitemtrackingprocess
+package utils
 
 import (
 	"context"
@@ -6,11 +6,10 @@ import (
 	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
-	"github.com/microsoft/terraform-provider-azuredevops/azuredevops/internal/utils"
 )
 
-// retryOnCondition executes f and retries if isRetryable returns true for the error.
-func retryOnCondition(ctx context.Context, timeout time.Duration, f func() error, isRetryable func(error) bool) error {
+// RetryOnCondition executes f and retries if isRetryable returns true for the error.
+func RetryOnCondition(ctx context.Context, timeout time.Duration, f func() error, isRetryable func(error) bool) error {
 	return retry.RetryContext(ctx, timeout, func() *retry.RetryError {
 		err := f()
 		if err == nil {
@@ -28,25 +27,25 @@ func retryOnCondition(ctx context.Context, timeout time.Duration, f func() error
 // eventual consistency issues where a resource cannot be modified yet because
 // dependent resources haven't been fully processed.
 func RetryOnUnexpectedException(ctx context.Context, timeout time.Duration, f func() error) error {
-	return retryOnCondition(ctx, timeout, f, func(err error) bool {
-		return utils.ResponseWasTypeKey(err, "UnexpectedException")
+	return RetryOnCondition(ctx, timeout, f, func(err error) bool {
+		return ResponseWasTypeKey(err, "UnexpectedException")
 	})
 }
 
-// retryOnNotFound retries the given function when Azure DevOps returns
+// RetryOnNotFound retries the given function when Azure DevOps returns
 // a 404 Not Found status code. This handles eventual consistency where
 // a newly created resource may not immediately be visible.
-func retryOnNotFound(ctx context.Context, timeout time.Duration, f func() error) error {
-	return retryOnCondition(ctx, timeout, f, func(err error) bool {
-		return utils.ResponseWasStatusCode(err, http.StatusNotFound)
+func RetryOnNotFound(ctx context.Context, timeout time.Duration, f func() error) error {
+	return RetryOnCondition(ctx, timeout, f, func(err error) bool {
+		return ResponseWasStatusCode(err, http.StatusNotFound)
 	})
 }
 
-// retryOnContributionNotFound retries the given function when Azure DevOps returns
+// RetryOnContributionNotFound retries the given function when Azure DevOps returns
 // a VS403120 error ("Contribution does not exist"). This handles eventual consistency
 // where a newly installed extension's contributions may not be immediately available.
-func retryOnContributionNotFound(ctx context.Context, timeout time.Duration, f func() error) error {
-	return retryOnCondition(ctx, timeout, f, func(err error) bool {
-		return utils.ResponseContainsStatusMessage(err, "VS403120")
+func RetryOnContributionNotFound(ctx context.Context, timeout time.Duration, f func() error) error {
+	return RetryOnCondition(ctx, timeout, f, func(err error) bool {
+		return ResponseContainsStatusMessage(err, "VS403120")
 	})
 }
